DBZ-8582 Support MariaDB 11.7+ vector data type in JDBCSinkConnector
