+++ debezium-connector-jdbc/src/main/java/io/debezium/connector/jdbc/dialect/maria/FloatVectorType.java
@@ -3,28 +3,28 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.debezium.connector.jdbc.dialect.mariadb;
+package io.debezium.connector.jdbc.dialect.maria;
 
 import java.util.Optional;
 
 import org.apache.kafka.connect.data.Schema;
 
-import io.debezium.connector.jdbc.type.debezium.AbstractDoubleVectorType;
+import io.debezium.connector.jdbc.type.debezium.AbstractFloatVectorType;
 import io.debezium.sink.column.ColumnDescriptor;
 
 /**
- * An implementation of {@link AbstractDoubleVectorType} for MySQL's {@code vector} data type.
+ * An implementation of {@link AbstractFloatVectorType} for Maria's {@code vector} data type.
  *
- * @author Chris Cranford
+ * @author Pranav Tiwari
  */
-public class DoubleVectorType extends AbstractDoubleVectorType {
+public class FloatVectorType extends AbstractFloatVectorType {
 
-    public static DoubleVectorType INSTANCE = new DoubleVectorType();
+    public static FloatVectorType INSTANCE = new FloatVectorType();
 
     @Override
     public String getTypeName(Schema schema, boolean isKey) {
         final Optional<String> size = getSourceColumnSize(schema);
-        return size.map(s -> String.format("vector(%s)", s)).orElse("vector(16383)");
+        return size.map(s -> String.format("vector(%s)", s)).orElse("vector(2048)");
     }
 
     @Override
